requirements-dev.txt
-r requirements.txt

factory-boy
pytest
pytest-django
